Less14/multipage_streamlit_teamplate/requirements.txt
streamlit
streamlit_option_menu
pandas
pillow
joblib
scikit-learn
streamlit-folium
